Fix CronJobrow.showLastFailure logic-typos, status/class ui improvs
When the oldest jobs of a cronjob failed but still were running on waiting
state(CreateContainerConfigError) and after fixing the cronjob definition the most
recents jobs succeed, cronJobRow threw errors because it couldn't get the correct
lastSuccedded/lastFailed attributes. It assumed all jobs of a cronjob have either
failed or succeeded.